+++ WIN3/NT/PRIVATE/SDKTOOLS/WPERF/DRAW.C
@@ -144,7 +144,7 @@ Revision History:
     // convert back to integer
     //
 
-    WindowsX = fWindowsX;
+    WindowsX = (int)fWindowsX;
 
     if (WindowsX == 0) {
         WindowsX = 1;
@@ -306,8 +306,8 @@ Revision History:
         fx2 = fx1 * 0.6666;
         fy1 = fx1 * 0.3333;
 
-        GraphHeight = fx2;
-        TextHeight  = fy1;
+        GraphHeight = (LONG)fx2;
+        TextHeight  = (LONG)fy1;
 
         if (TextHeight > 20) {
             GraphHeight += TextHeight-20;
